bsps/sparc: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This addresses warnings that only occurred on 64-bit targets.  For the
ones which only appeared on 64-bit targets, the cause was frequently
a mismatch when comparing a combination off_t, ssize_t, and int.

+++ bsps/sparc/erc32/console/erc32_console.c
@@ -292,7 +292,7 @@ static ssize_t erc32_console_write_support_polled(
   size_t      len
 )
 {
-  int nwrite = 0;
+  size_t nwrite = 0;
 
   while (nwrite < len) {
     console_outbyte_polled( minor, *buf++ );
